chore: update synthetics_multilocation_alert to use monitor GUID
